+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ### X.X.X (coming soon)
+
+Bugfixes:
+  - LokiJS: Do not call `saveDatabase()` when no persistence adapter is given.
+
 ### 10.5.3 (19 November 2021)
 
 Bugfixes: